requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1720453481653/work/certifi
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna @ file:///croot/idna_1714398848350/work
importlib_resources==6.4.0
Jinja2 @ file:///croot/jinja2_1716993405101/work
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.0.3
pillow @ file:///croot/pillow_1714398848491/work
platformdirs==4.2.2
pluggy==1.5.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pyproject-api==1.7.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1716902831423/work
rich==13.7.1
scipy==1.10.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy @ file:///croot/sympy_1701397643339/work
tomli==2.0.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tox==4.16.0
tqdm==4.66.4
triton==2.3.1
typer==0.12.3
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.54
ultralytics-thop==2.0.0
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
websockets==12.0
zipp==3.19.2
